Moved UploadControllerHelper to repository folder
