+++ repository/UploadControllerHelper.js
@@ -1,5 +1,5 @@
-const ImageModel = require("../../models/Image");
-const CategoryControllerHelper = require("../../repository/CategoryRepository");
+const ImageModel = require("../models/Image");
+const CategoryControllerHelper = require("./CategoryRepository");
 
 class UploadControllerHelper {
    async SaveNewImage(image, categoryTitle, newImages, description) {
